A2模块分数的计算，代码初步完成

+++ RuoYi-Vue-master/ruoyi-erock/src/main/java/com/ruoyi/core/service/SelectUser.java
@@ -46,7 +46,7 @@ public interface SelectUser<T extends BaseEntity>  {
 //    A1学习次数业务接口
     public Double A1calculationTimes(int number, int stuNumber);
 //    A2学习评价业务接口
-    public int A2calculationTimes(int excellent, int ordinary);
+    public Double A2calculationTimes(int excellent, int ordinary);
